+++ backend/core/types/src/backend-swagger.ts
@@ -702,13 +702,19 @@ export class AuthService {
   /**
    * Token
    */
-  token(options: IRequestOptions = {}): Promise<LoginResponse> {
+  token(
+    params: {
+      /** requestBody */
+      body?: Token
+    } = {} as any,
+    options: IRequestOptions = {}
+  ): Promise<LoginResponse> {
     return new Promise((resolve, reject) => {
       let url = basePath + "/auth/token"
 
       const configs: IRequestConfig = getConfigs("post", "application/json", url, options)
 
-      let data = null
+      let data = params.body
 
       configs.data = data
       axios(configs, resolve, reject)
@@ -3909,6 +3915,8 @@ export interface LoginResponse {
   accessToken: string
 }
 
+export interface Token {}
+
 export interface RequestMfaCode {
   /**  */
   email: string
@@ -4435,6 +4443,9 @@ export interface JurisdictionCreate {
   emailFromAddress: string
 
   /**  */
+  rentalAssistanceDefault: string
+
+  /**  */
   programs: Id[]
 
   /**  */
@@ -4468,6 +4479,9 @@ export interface JurisdictionUpdate {
 
   /**  */
   emailFromAddress: string
+
+  /**  */
+  rentalAssistanceDefault: string
 
   /**  */
   programs: Id[]