+++ sites/public/cypress/integration/pages/application/review/confirmation.spec.ts
@@ -5,14 +5,9 @@ describe("applications/review/confirmation", function () {
     cy.visit(route)
   })
 
-  it("should redirect to create account page on click", function () {
-    cy.getByTestId("app-confirmation-create-account").click()
-    cy.location("pathname").should("include", "/create-account")
-  })
-
   it("should redirect to home page on click", function () {
-    cy.getByTestId("app-confirmation-done").click()
-    cy.location("pathname").should("equals", "/")
+    cy.getByTestId("app-confirmation-browse").click()
+    cy.location("pathname").should("equals", "/listings")
   })
 
   it("should redirect to create account page", function () {